+++ src/main/java/com/booleanuk/api/cinema/controllers/MovieController.java
@@ -5,6 +5,7 @@ import com.booleanuk.api.cinema.models.Movie;
 import com.booleanuk.api.cinema.models.Screening;
 import com.booleanuk.api.cinema.repository.MovieRepository;
 import com.booleanuk.api.cinema.repository.ScreeningRepository;
+import com.booleanuk.api.cinema.responses.ErrorResponse;
 import com.booleanuk.api.cinema.responses.MovieListResponse;
 import com.booleanuk.api.cinema.responses.MovieResponse;
 import com.booleanuk.api.cinema.responses.Response;
@@ -41,6 +42,12 @@ public class MovieController {
 
     @PostMapping
     public ResponseEntity<Response<?>> createMovie(@RequestBody Movie movie){
+        if (movie.getTitle() == null || movie.getRating() == null || movie.getDescription() == null || movie.getRuntimeMins() <= 0){
+            ErrorResponse error = new ErrorResponse();
+            error.set("Could not create a new movie, please check all fields are correct");
+            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+        }
+
         movie.setCreatedTime(LocalDateTime.now());
         movie.setUpdatedTime(LocalDateTime.now());
 
@@ -52,7 +59,18 @@ public class MovieController {
 
     @PutMapping("{id}")
     public ResponseEntity<Response<?>> updateMovie(@PathVariable int id, @RequestBody Movie movie){
-        Movie movieToUpdate = this.movieRepository.findById(id).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "Not Found"));
+        if (movie.getTitle() == null || movie.getRating() == null || movie.getDescription() == null || movie.getRuntimeMins() <= 0){
+            ErrorResponse error = new ErrorResponse();
+            error.set("Could not create a new movie, please check all fields are correct");
+            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+        }
+
+        Movie movieToUpdate = this.movieRepository.findById(id).orElse(null);
+        if (movieToUpdate == null){
+            ErrorResponse error = new ErrorResponse();
+            error.set("No movie with that id found");
+            return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+        }
 
         movieToUpdate.setTitle(movie.getTitle());
         movieToUpdate.setRating(movie.getRating());
@@ -67,7 +85,13 @@ public class MovieController {
 
     @DeleteMapping("{id}")
     public ResponseEntity<Response<?>> deleteMovie(@PathVariable int id){
-        Movie movieToDelete = this.movieRepository.findById(id).orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "Not Found"));
+        Movie movieToDelete = this.movieRepository.findById(id).orElse(null);
+        if (movieToDelete == null){
+            ErrorResponse error = new ErrorResponse();
+            error.set("No movie with that id found");
+
+            return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+        }
 
         this.movieRepository.delete(movieToDelete);
         this.movieResponse.set(movieToDelete);
